[Refactor] component 분리 및 로그인 버튼 로직 변경 - component 분리: Aside, Feed, InfoList, UsersStory - 로그인 버튼 로직 변경: 로그인 버튼 활성화 되면 바로 메인으로 이동 되도록 변경
