+++ src/pages/heewonKang/Main/InfoList.js
@@ -1,0 +1,29 @@
+export default function InfoList() {
+  const footerList = [
+    { id: 1, content: '소개' },
+    { id: 2, content: '도움말' },
+    { id: 3, content: '홍보 센터' },
+    { id: 4, content: 'API' },
+    { id: 5, content: '채용 정보' },
+    { id: 6, content: '개인정보처리방침' },
+    { id: 7, content: '약관' },
+    { id: 8, content: '위치' },
+    { id: 9, content: '인기 계정' },
+    { id: 10, content: '해시태그' },
+    { id: 11, content: '언어' },
+  ];
+  return (
+    <footer>
+      <div className="footer_link_list">
+        {footerList.map(item => {
+          return (
+            <a href="#!" key={item.id}>
+              {item.content}
+            </a>
+          );
+        })}
+      </div>
+      <address>© 2021 WESTAGRAM FROM WETA</address>
+    </footer>
+  );
+}
