+++ src/pages/heewonKang/Main/UsersStory.js
@@ -1,0 +1,116 @@
+export default function UsersStory() {
+  return (
+    <div className="users_story">
+      <div className="users_story_content">
+        <div>
+          <div className="arrows">
+            <i className="fas fa-chevron-left" />
+            <i className="fas fa-chevron-right" />
+          </div>
+          <ul className="story">
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>kangheewon</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>kangminhee</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+            <li>
+              <div>
+                <img
+                  className="circle_profile_img"
+                  src="https://images.unsplash.com/photo-1639502496516-95531e23e304?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1365&q=80"
+                  alt="story_user_profile_img"
+                />
+              </div>
+              <em>dlwlrma_</em>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+  );
+}
